SafetyNet 101
What is SafetyNet, how does it work, and why should it be built? How will SafetyNet `actually` be used?
This is for Aisha, the Cathouse Crew, the City of Las Vegas Metro Police Department, especially SouthEast Command, especially King and John sectors (I think).
Everything starts locally, and this is for the hometown team.